requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cog==0.9.9
cachetools==5.5.0
colorama==0.4.6
# removed conflicting 'discord' package; using discord.py instead
discord.py[voice]==2.4.0
dnspython==2.6.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
fuzzywuzzy==0.18.0
huggingface-hub==0.23.0
idna==3.7
motor==3.4.0
multidict==6.0.5
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
packaging==24.0
pymongo==4.7.3
python-Levenshtein==0.25.1
python-dotenv==1.0.1
pytesseract==0.3.10
PyYAML==6.0.1
PyNaCl==1.5.0
rapidfuzz==3.9.5
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.40.1
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
